+++ forms-flow-nav/src/endpoints/index.js
@@ -8,6 +8,8 @@ const API = {
     USER_PROFILE_UPDATE: `${WEB_BASE_URL}/user/<user_id>/profile`,
     GET_PERMISSIONS: `${WEB_BASE_URL}/roles/permissions`,
     USER_LOGIN_DETAILS: (userId) => `${WEB_BASE_URL}/user/${userId}/login-details`,
+    ONBOARDING_USER_ROLE: `${WEB_BASE_URL}/user/info`,
+    CHECKLIST: `${WEB_BASE_URL}/user/checklist`,
 }
 
 export default API;